+++ packages/flow-design/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coding-flow/flow-design
 
+## 0.1.9
+
+### Patch Changes
+
+- @coding-flow/flow-core@0.1.9
+- @coding-flow/flow-icons@0.1.9
+- @coding-flow/flow-pc-ui@0.1.9
+- @coding-flow/flow-types@0.1.9
+
 ## 0.1.8
 
 ### Patch Changes