+++ packages/flow-pc/flow-pc-approval/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @coding-flow/flow-pc-approval
 
+## 0.1.9
+
+### Patch Changes
+
+- @coding-flow/flow-approval-presenter@0.1.9
+- @coding-flow/flow-core@0.1.9
+- @coding-flow/flow-icons@0.1.9
+- @coding-flow/flow-pc-form@0.1.9
+- @coding-flow/flow-pc-ui@0.1.9
+- @coding-flow/flow-types@0.1.9
+
 ## 0.1.8
 
 ### Patch Changes