+++ packages/flow-pc/flow-pc-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @coding-flow/flow-pc-ui
 
+## 0.1.9
+
+### Patch Changes
+
+- @coding-flow/flow-core@0.1.9
+
 ## 0.1.8
 
 ### Patch Changes