+++ packages/flow-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @coding-flow/flow-types
 
+## 0.1.9
+
+### Patch Changes
+
+- @coding-flow/flow-core@0.1.9
+
 ## 0.1.8
 
 ### Patch Changes